+++ tests/test_cloud_storage.py
@@ -210,10 +210,20 @@ class TestGcsBucket:
             bucket_folder=request.param,
         )
 
-    def test_list_folders(self, gcs_bucket_with_bucket_folder):
+    def test_list_folders_with_root_only(self, gcs_bucket_with_bucket_folder):
         blobs = gcs_bucket_with_bucket_folder.list_folders()
+        assert len(blobs) == 2
+        assert blobs[0] == "base_folder/sub_folder"
+        assert blobs[1] == "base_folder"
+
+
+
+    def test_list_folders_with_sub_folders(self, gcs_bucket_with_bucket_folder):
+        blobs = gcs_bucket_with_bucket_folder.list_folders("base_folder")
         assert len(blobs) == 1
-        assert blobs[0] == "sub_folder"
+        assert blobs[0] == "base_folder/sub_folder"
+
+
     def test_list_blobs(self, gcs_bucket_no_bucket_folder):
         blobs = gcs_bucket_no_bucket_folder.list_blobs(folder="base_folder/")
         assert len(blobs) == 2